Refactor & Systematize a few style related elements (CSS/Layout)
This takes a little bit from Next.js learner documentation as well as
Refactoring UI in a step toward "systematizing things" to reduce decision
time over all. I wanted to slean things up a bit for the existing page
prior to adding in the podcast list page, and this is my first pass at that.